+++ Source/Tools/RampGenerator/RampGenerator.cpp
@@ -337,7 +337,7 @@ void WriteIES(unsigned char* data, unsigned width, unsigned height, PODVector<fl
     // otherwise the space used would fit the light's traits and potentially incude a lot of wasted black space
     const float angularFactor = 90.0f;
     const float fraction = angularFactor / ((float)width);
-    ::memset(data, 0, width * height);
+    ::memset(data, 0, (size_t)width * height);
 
     for (unsigned y = 0; y < height; ++y)
     {
