+++ Source/Tools/SpritePacker/SpritePacker.cpp
@@ -338,7 +338,7 @@ void Run(Vector<String>& arguments)
     spriteSheetImage.SetSize(packedWidth, packedHeight, 4);
 
     // zero out image
-    spriteSheetImage.SetData((unsigned char*)calloc(sizeof(unsigned char), packedWidth * packedHeight * 4));
+    spriteSheetImage.SetData((unsigned char*)calloc(sizeof(unsigned char), (size_t)packedWidth * packedHeight * 4));
 
     XMLFile xml(context);
     XMLElement root = xml.CreateRoot("TextureAtlas");
